Bet services on angular

+++ angular-client-spa/src/app/admin/content/admin-matches.component.ts
@@ -115,53 +115,57 @@ export class AdminMatchesComponent implements OnInit {
     newLOLMatch(match: EgamesMatch) {
         match.type = 'LOL';
         match.time = match.time.concat(':00');
-         if (!this.checkTeams(match.homeTeam, match.visitingTeam) &&
+        if (!this.checkTeams(match.homeTeam, match.visitingTeam) &&
             !this.checkQuotas(match.quotaHomeVictory, match.quotaVisitingVictory, 0,
-            match.quotaHomeFirstBlood, match.quotaVisitingFirstBlood)) {
-                this.matchService.newEgamesMatch(match)
-                    .then(() => {
-                        this.getLolMatches();
-                        this.addAlert('Se ha añadido un nuevo partido de LOL');
-                    });
-            } else {
-                this.addBadAlert('Comprueba que los equipos no sean iguales y que cada para de quotas sumen 110');
-            }
+                match.quotaHomeFirstBlood, match.quotaVisitingFirstBlood)) {
+            this.matchService.newEgamesMatch(match)
+                .then(() => {
+                    this.getLolMatches();
+                    this.addAlert('Se ha añadido un nuevo partido de LOL');
+                });
+        } else {
+            this.addBadAlert('Comprueba que los equipos no sean iguales y que cada para de quotas sumen 110');
+        }
     }
 
     newCSMatch(match: EgamesMatch) {
         match.type = 'CS-GO';
         match.time = match.time.concat(':00');
-         if (!this.checkTeams(match.homeTeam, match.visitingTeam) &&
+        if (!this.checkTeams(match.homeTeam, match.visitingTeam) &&
             !this.checkQuotas(match.quotaHomeVictory, match.quotaVisitingVictory, 0,
-            match.quotaHomeFirstBlood, match.quotaVisitingFirstBlood)) {
-                this.matchService.newEgamesMatch(match)
-                    .then(() => {
-                        this.getCsMatches();
-                        this.addAlert('Se ha añadido un nuevo partido de CS-GO');
-                    });
-            } else {
-                this.addBadAlert('Comprueba que los equipos no sean iguales y que cada para de quotas sumen 110');
-            }
+                match.quotaHomeFirstBlood, match.quotaVisitingFirstBlood)) {
+            this.matchService.newEgamesMatch(match)
+                .then(() => {
+                    this.getCsMatches();
+                    this.addAlert('Se ha añadido un nuevo partido de CS-GO');
+                });
+        } else {
+            this.addBadAlert('Comprueba que los equipos no sean iguales y que cada para de quotas sumen 110');
+        }
     }
 
     removeSportMatch(match: SportMatch) {
         const confirmationMessage = confirm('¿Estás seguro de que quieres borrar el equipo con id '
             + match.id + '?');
-        this.matchService.removeSportMatchResult(match.id)
-            .then(() => {
-                this.getFootballMatches();
-                this.getBasketballMatches();
-            });
+        if (confirmationMessage) {
+            this.matchService.removeSportMatchResult(match.id)
+                .then(() => {
+                    this.getFootballMatches();
+                    this.getBasketballMatches();
+                });
+        }
     }
 
     removeEgamesMatch(match: EgamesMatch) {
         const confirmationMessage = confirm('¿Estás seguro de que quieres borrar el equipo con id '
             + match.id + '?');
-        this.matchService.removeEgamesMatchResult(match.id)
-            .then(() => {
-                this.getLolMatches();
-                this.getCsMatches();
-            });
+        if (confirmationMessage) {
+            this.matchService.removeEgamesMatchResult(match.id)
+                .then(() => {
+                    this.getLolMatches();
+                    this.getCsMatches();
+                });
+        }
     }
 
     getFootballMatches() {
